refactor: move PT auth ITs into auth package
So they match the identity package filters as defined in the `identity-packages`
of the pom.xml.
