+++ qa/acceptance-tests/src/test/java/io/camunda/it/auth/MultiPhysicalTenantAuthorizationIT.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.it.physicaltenant;
+package io.camunda.it.auth;
 
 import static io.camunda.client.api.search.enums.OwnerType.GROUP;
 import static io.camunda.client.api.search.enums.OwnerType.MAPPING_RULE;
