+++ qa/acceptance-tests/src/test/java/io/camunda/it/auth/PhysicalTenantLogicalTenantScopingIT.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.it.physicaltenant;
+package io.camunda.it.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
